Move SetDevelocityProjectId to plugins package
Aligns with the other Develocity recipes (AddDevelocityGradlePlugin,
RemoveDevelocityConfiguration, MigrateGradleEnterpriseToDevelocity),
which live in org.openrewrite.gradle.plugins.
